requirements.txt
langchain==0.3.25
langchain-openai==0.3.16
langchain-community==0.3.24
langgraph==0.2.73
langfuse==2.60.2
openai==1.82.0
faiss-cpu==1.11.0
tiktoken==0.9.0
python-dotenv==1.1.0
jupyter==1.1.1
ipykernel==6.29.5
pypdf==5.5.0

# Testing
pytest==8.3.5
pytest-cov==6.1.0